+++ packages/dumb/src/components/tooltips/physical-tooltip/physical-tooltip.tsx
@@ -41,7 +41,13 @@ import { formatInputTime } from '../../../utils/format/format.util'
 import colors from '../../../styles/colors.css'
 import { getDateTitleForBaseDatum } from '../../../utils/tooltip/tooltip.util'
 import { convertValueToMinutes } from '../../../utils/datetime/datetime.util'
-import { DurationUnit, type DurationValue, type PhysicalActivity, type TimePrefs } from 'medical-domain'
+import {
+  DurationUnit,
+  type DurationValue,
+  type PhysicalActivity,
+  PhysicalActivityName,
+  type TimePrefs
+} from 'medical-domain'
 import { useTranslation } from 'react-i18next'
 import { TooltipLine } from '../common/tooltip-line/tooltip-line'
 
@@ -54,7 +60,7 @@ interface PhysicalTooltipProps {
 
 export const PhysicalTooltip: FunctionComponent<PhysicalTooltipProps> = (props) => {
   const { physicalActivity, position, side, timePrefs } = props
-  const { t } = useTranslation('main')
+  const { t } = useTranslation()
 
   const getDurationInMinutes = (): DurationValue => {
     const units = physicalActivity?.duration?.units
@@ -68,6 +74,14 @@ export const PhysicalTooltip: FunctionComponent<PhysicalTooltipProps> = (props) 
   }
 
   const duration = getDurationInMinutes()
+
+  const getDisplayName = (name: string): string => {
+    const nameUppercase = name.toUpperCase()
+    if (name && Object.values(PhysicalActivityName).includes(nameUppercase as PhysicalActivityName)) {
+      return t(`params|${nameUppercase}`)
+    }
+    return t(`params|${PhysicalActivityName.AerobicDefault}`)
+  }
 
   return (
     <Tooltip
@@ -83,6 +97,10 @@ export const PhysicalTooltip: FunctionComponent<PhysicalTooltipProps> = (props) 
       content={
         <div className={commonStyles.containerFlex}>
           <TooltipLine label={t('Physical Activity')} isBold />
+          {
+            physicalActivity.name !== "" &&
+            <TooltipLine label={t('Name')} value={getDisplayName(physicalActivity.name)} />
+          }
           <TooltipLine label={t('Intensity')} value={t(`${physicalActivity.reportedIntensity}-pa`)} />
           <TooltipLine label={t('Duration')} value={`${duration.value} ${t(duration.units)}`} />
           {