+++ packages/medical-domain/test/repositories/medical/medical-data.service.test.ts
@@ -467,7 +467,8 @@ const testData = {
       "epochEnd": 1698567720000,
       "guid": "physicalActivity-16974144000000003",
       "reportedIntensity": "low",
-      "inputTime": "2021-01-15T08:21:47Z"
+      "inputTime": "2021-01-15T08:21:47Z",
+      "name": ""
     },
     {
       "id": "318b802601eed0eeb8076a4ed4aa9c5b",
@@ -485,7 +486,8 @@ const testData = {
       "epochEnd": 1698395050000,
       "guid": "physicalActivity-16974144000000002",
       "reportedIntensity": "low",
-      "inputTime": "2021-01-13T08:02:18Z"
+      "inputTime": "2021-01-13T08:02:18Z",
+      "name": ""
     }
   ],
   pumpSettings: [